cosmosdb

+++ src/main/java/com/testcontainers/catalog/domain/models/Product.java
@@ -3,4 +3,4 @@ package com.testcontainers.catalog.domain.models;
 import java.math.BigDecimal;
 
 public record Product(
-        Long id, String code, String name, String description, String imageUrl, BigDecimal price, boolean available) {}
+        Long id, String code, String name, String description, String imageUrl, Double price, boolean available) {}
